test(research): extract span caption fixtures
